+++ frontend/fingertips-frontend/playwright/tests/fully_integrated_e2e_tests/core_journey_config.ts
@@ -1,0 +1,118 @@
+import {
+  AreaMode,
+  IndicatorMode,
+  SearchMode,
+  TestParams,
+} from '@/playwright/testHelpers';
+
+export const coreTestJourneys: TestParams[] = [
+  {
+    indicatorMode: IndicatorMode.ONE_INDICATOR,
+    areaMode: AreaMode.ENGLAND_AREA,
+    searchMode: SearchMode.ONLY_SUBJECT,
+    subjectSearchTerm: '22401', // test searching for a specific indicatorID
+    indicatorsToSelect: [
+      {
+        indicatorID: '22401',
+        knownTrend: 'Decreasing and getting better',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.ONE_INDICATOR,
+    areaMode: AreaMode.ONE_AREA,
+    searchMode: SearchMode.BOTH_SUBJECT_AND_AREA,
+    subjectSearchTerm: 'emergency',
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+        knownTrend: 'No recent trend data available',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.ONE_INDICATOR,
+    areaMode: AreaMode.THREE_PLUS_AREAS,
+    searchMode: SearchMode.ONLY_SUBJECT,
+    subjectSearchTerm: 'emergency',
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.ONE_INDICATOR,
+    areaMode: AreaMode.ALL_AREAS_IN_A_GROUP,
+    searchMode: SearchMode.ONLY_SUBJECT,
+    subjectSearchTerm: 'emergency',
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+        knownTrend: 'No recent trend data available',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.TWO_INDICATORS,
+    areaMode: AreaMode.ENGLAND_AREA,
+    searchMode: SearchMode.ONLY_SUBJECT,
+    subjectSearchTerm: 'emergency',
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+        knownTrend: 'No recent trend data available',
+      },
+      {
+        indicatorID: '22401',
+        knownTrend: 'Decreasing and getting better',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.TWO_INDICATORS,
+    areaMode: AreaMode.ALL_AREAS_IN_A_GROUP,
+    searchMode: SearchMode.ONLY_SUBJECT,
+    subjectSearchTerm: 'emergency',
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+        knownTrend: 'No recent trend data available',
+      },
+      {
+        indicatorID: '22401',
+        knownTrend: 'Decreasing and getting better',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.TWO_INDICATORS,
+    areaMode: AreaMode.THREE_PLUS_AREAS,
+    searchMode: SearchMode.ONLY_AREA, // therefore no subject search term required
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+      },
+      {
+        indicatorID: '22401',
+      },
+    ],
+  },
+  {
+    indicatorMode: IndicatorMode.THREE_PLUS_INDICATORS,
+    areaMode: AreaMode.TWO_AREAS,
+    searchMode: SearchMode.ONLY_SUBJECT,
+    subjectSearchTerm: 'hospital', // a different subject search term is required that returns enough search results allowing for three indicators to be selected
+    indicatorsToSelect: [
+      {
+        indicatorID: '41101',
+      },
+      {
+        indicatorID: '22401',
+      },
+      {
+        indicatorID: '91894',
+      },
+    ],
+  },
+];
